General improvements to the web app Updated both client versions to display the VRAM amount in either MB or GB Fixed the backend GraphQL server to accept MB VRAM values and memory speeds of less than 1Gbps Fixed the server port routing for both clients in the Docker compose orchestrations, apply for both development and production versions Small fixes and improvements to the doc file

+++ server/src/schemas/schemas.js
@@ -14,7 +14,7 @@ const typeDefs = gql`
     cores: Int!
     tmus: Int!
     rops: Int!
-    vram: Int!
+    vram: Float!
     bus: Int!
     memtype: String!
     baseclock: Int!
@@ -30,7 +30,7 @@ const typeDefs = gql`
       cores: Int!
       tmus: Int!
       rops: Int!
-      vram: Int!
+      vram: Float!
       bus: Int!
       memtype: String!
       baseclock: Int!
